Fix wardriving BLE reconnect-stall coexistence bug; add GPS antenna test tooling; consolidate docs
Wardriving: BLE reconnect stalls were caused by wardriving's Wi-Fi source
starving the shared 2.4GHz radio while a BLE reconnect was in flight -
now throttled to a 400ms gap during reconnect, restored to the configured
cadence on successful connect. Also raise ble_window_ms's default from
30ms to 100ms (~20% duty): the prior 6% duty's 30ms window/470ms idle gap
gave short test runs a real chance of missing nearby devices by bad luck.

Add a persistent GPS antenna smoke-test firmware (esp32/gps_antenna_test)
and PowerShell test harness (tools/test_gps_antenna.ps1) to validate the
GPS module ahead of the GPS capability work.

Consolidate documentation: fold FINDINGS_BACKLOG.md/OPTIMIZATION.md into
BACKLOG.md, add cross-model findings appendices under
flipper_over_ble_findings/, rewrite README.md to match current protocol
and capability state, and update PLAN.md/SESSION_MEMORY.md/PROJECT_HISTORY.md
accordingly.

+++ flipper/wardriving_csv.h
@@ -111,7 +111,7 @@ uint32_t feb_wardriving_backdate_first_seen(
    gets deduplicated. */
 #define FEB_WARDRIVING_DEDUP_CAPACITY 256u
 #define FEB_WARDRIVING_DEDUP_RSSI_IMPROVE_DB 6
-#define FEB_WARDRIVING_DEDUP_MOVE_METERS 30.0
+#define FEB_WARDRIVING_DEDUP_MOVE_METERS ((double)30.0)
 
 typedef struct {
     uint8_t address[6];
